+++ package.js
@@ -2,7 +2,7 @@
 Package.describe({
   name: 'reywood:publish-composite',
   summary: 'Publish a set of related documents from multiple collections with a reactive join.',
-  version: '1.8.5-beta.2',
+  version: '1.8.6',
   git: 'https://github.com/Meteor-Community-Packages/meteor-publish-composite'
 })
 
@@ -11,14 +11,15 @@ Npm.depends({
 })
 
 Package.onUse((api) => {
-  api.versionsFrom(['1.8.3', '2.8.1', '3.0-alpha.19'])
+  api.versionsFrom(['1.8.3', '2.8.1', '3.0-beta.0'])
   api.use([
     'check',
     'ecmascript',
     'modules',
-    'logging'
+    'logging',
+    'ddp'
   ], ['server'])
-  api.use('zodern:types@1.0.10')
+  api.use('zodern:types@1.0.11')
   api.mainModule('lib/publish_composite.js', 'server')
   api.addFiles([
     'lib/doc_ref_counter.js',